batch/requirements.txt
mysql-connector-python
python-dotenv
openai
requests
beautifulsoup4
lxml
feedparser